Complete SDHPC use case section
Renamed section to highlight SDHPC use for data and
cluster migration.
Describe the use case that drove the developement of the
SDHPC approach.
Provide specfics to the the migraiton workflow, hosting
approach and use of the CICD pipelines.

Remove notes from start of section.
Move notes to section three.
Add AI Infrastructure keyword.