Comment out subsections on build and deploy
These subsections are too detailed and repeat information
no provided in the main body of the CICD section.
Comment out the material before committing to deletion
to see if it's needed for the narrative.
Add the note about versioned builds to the build
commentary since that seems useful to know.