Introduce new section for SSH throughput performance analysis
Create a distinct section for the ssh performance analysis.
This focuses the performance narrative on the tested scenarios and
keeps it seperate from the discussion of the migration scenario.
Rewrote the conclusions based on the new results in the table.
Providing a full breakdown of the data in the table.

Refined the migration section discussion in like of the new section.
Moved the commentary on RCS deployment and dev support to the migration
section.
Tuned the wording in the migration section to improve flow of the
narrative.